backend/social/requirements.txt
fastapi
uvicorn
pydantic-settings
python-jose[cryptography]
azure-cosmos